refactor(storage-core): move storage files out of @bitwarden/common
