+++ libs/common/src/platform/enums/html-storage-location.enum.ts
@@ -1,7 +1,1 @@
-// FIXME: update to use a const object instead of a typescript enum
-// eslint-disable-next-line @bitwarden/platform/no-enums
-export enum HtmlStorageLocation {
-  Local = "local",
-  Memory = "memory",
-  Session = "session",
-}
+export { HtmlStorageLocation } from "@bitwarden/storage-core";
