+++ libs/common/src/platform/enums/storage-location.enum.ts
@@ -1,7 +1,1 @@
-// FIXME: update to use a const object instead of a typescript enum
-// eslint-disable-next-line @bitwarden/platform/no-enums
-export enum StorageLocation {
-  Both = "both",
-  Disk = "disk",
-  Memory = "memory",
-}
+export { StorageLocationEnum as StorageLocation } from "@bitwarden/storage-core";
