+++ libs/storage-core/src/index.ts
@@ -1,0 +1,9 @@
+export * from "./client-locations";
+export * from "./html-storage-location.enum";
+export * from "./memory-storage.service";
+export * from "./serialized-memory-storage.service";
+export * from "./storage-location";
+export * from "./storage-location.enum";
+export * from "./storage-options";
+export * from "./storage-service.provider";
+export * from "./storage.service";
