+++ libs/storage-core/src/storage-options.ts
@@ -1,0 +1,10 @@
+import { HtmlStorageLocation } from "./html-storage-location.enum";
+import { StorageLocationEnum as StorageLocation } from "./storage-location.enum";
+
+export type StorageOptions = {
+  storageLocation?: StorageLocation;
+  useSecureStorage?: boolean;
+  userId?: string;
+  htmlStorageLocation?: HtmlStorageLocation;
+  keySuffix?: string;
+};
